+++ packable/packable/src/unpacker/mod.rs
@@ -6,10 +6,12 @@
 //! The [`Unpacker`] trait represents types that can be used to read bytes from it. It can be thought as a `no_std`
 //! friendly alternative to the [`Read`](std::io::Read) trait.
 
+mod counter;
 #[cfg(feature = "io")]
 mod io;
 mod slice;
 
+pub use counter::CounterUnpacker;
 #[cfg(feature = "io")]
 pub use io::IoUnpacker;
 
@@ -29,5 +31,19 @@ pub trait Unpacker: Sized {
     /// it is allowed to return `Ok(())` in any other case.
     fn ensure_bytes(&self, _len: usize) -> Result<(), Self::Error> {
         Ok(())
+    }
+}
+
+impl<U: Unpacker + ?Sized> Unpacker for &mut U {
+    type Error = U::Error;
+
+    #[inline]
+    fn unpack_bytes<B: AsMut<[u8]>>(&mut self, bytes: B) -> Result<(), Self::Error> {
+        U::unpack_bytes(*self, bytes)
+    }
+
+    #[inline]
+    fn ensure_bytes(&self, len: usize) -> Result<(), Self::Error> {
+        U::ensure_bytes(*self, len)
     }
 }